requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.0
gast==0.5.5
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h5py==3.11.0
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.4.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-text-splitters==0.2.2
langsmith==0.1.82
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.5
packaging==23.2
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.1.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
streamlit==1.36.0
tenacity==8.4.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.16.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.41.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
